+++ src/main/java/dev/shoxruhjon/services/IWalletService.java
@@ -2,8 +2,10 @@ package dev.shoxruhjon.services;
 
 import dev.shoxruhjon.models.User;
 
+import java.math.BigDecimal;
+
 public interface IWalletService {
-    void topUp(User user, double amount);
+    void topUp(User user, BigDecimal amount);
     void printWallet(User user);
     void printHistory(String userId);
 }
